Moved tests for clean string attrs to write utils
